[cms] Add check for only 1 invoice per month

+++ politeiawww/cmd/politeiawwwcli/commands/newinvoice.go
@@ -161,8 +161,8 @@ func (cmd *NewInvoiceCmd) Execute(args []string) error {
 		Files:     files,
 		PublicKey: hex.EncodeToString(cfg.Identity.Public.Key[:]),
 		Signature: sig,
-		Month:     uint16(month),
-		Year:      uint16(year),
+		Month:     month,
+		Year:      year,
 	}
 
 	// Print request details
